python/requirements.txt
anthropic==0.34.0
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
